+++ packages/core/components/card/index.mdx
@@ -7,7 +7,7 @@ route: /components/card
 import { Playground, Props } from 'docz'
 import Card from './'
 import { Density } from "../../common/density";
-import { Header, Text, Button } from "../../"
+import { Heading, Text, Button } from "../../"
 
 # Card
 
@@ -42,7 +42,7 @@ Composes a [`Box`](/typography/box) with preset styles.
 
 <Playground>
   <Card density="comfortable">
-    <Header tag="h4">Hello world</Header>
+    <Heading tag="h4">Hello world</Heading>
     <Text tag="p" mb="30px">Lorem ipsum dolor, sit amet consectetur adipisicing elit. Id magni repellendus illum quasi, officiis assumenda facere inventore accusamus maiores soluta, qui quibusdam et, dolorum tempora ab a unde porro eligendi!</Text>
     <Button appearance="prominent">Continue</Button>
     <Button appearance="minimal">Cancel</Button>
@@ -58,9 +58,9 @@ Composes a [`Box`](/typography/box) with preset styles.
     bg="primaryTint"
     borderColor="primaryShade"
   >
-    <Header tag="h4">
+    <Heading tag="h4">
       Congratulations!
-    </Header>
+    </Heading>
     <Text tag="p" mb="30px">
       Lorem ipsum dolor, sit amet consectetur adipisicing elit. Id magni repellendus illum quasi, officiis assumenda facere inventore accusamus maiores soluta, qui quibusdam et, dolorum tempora ab a unde porro eligendi!
     </Text>
@@ -107,5 +107,5 @@ Composes a [`Box`](/typography/box) with preset styles.
 }
 ```
 
-- [Default theme](https://github.com/raster-foundry/blasterjs/tree/master/packages/core/theme/components/header)
+- [Default theme](https://github.com/raster-foundry/blasterjs/tree/master/packages/core/theme/components/heading)
 - [Learn more about themeing](#)
