+++ packages/core/components/heading/index.mdx
@@ -1,43 +1,43 @@
 ---
-name: Header
+name: Heading
 menu: Typography
-route: /typography/header
+route: /typography/heading
 ---
 
 import { Playground, Props } from 'docz'
-import Header from './'
+import Heading from './'
 
-# Header
+# Heading
 
 ## Import
-`import { Header } from '@blasterjs/core'`
+`import { Heading } from '@blasterjs/core'`
 
 ## Basic Usage
 
 Composes [`Text`](/typography/text) and renders an `<h1>` by default.
 
 <Playground>
-  <Header>Header 1</Header>
+  <Heading>Heading 1</Heading>
 </Playground>
 
-## Changing header tag
+## Changing heading tag
 
 Use the `tag` prop to change from the default `h1` element to `h2`, `h3`, `h4`, `h5`, or `h6`.
 
 <Playground>
-  <Header tag="h1">Header 1</Header>
-  <Header tag="h2">Header 2</Header>
-  <Header tag="h3">Header 3</Header>
-  <Header tag="h4">Header 4</Header>
-  <Header tag="h5">Header 5</Header>
-  <Header tag="h6">Header 6</Header>
+  <Heading tag="h1">Heading 1</Heading>
+  <Heading tag="h2">Heading 2</Heading>
+  <Heading tag="h3">Heading 3</Heading>
+  <Heading tag="h4">Heading 4</Heading>
+  <Heading tag="h5">Heading 5</Heading>
+  <Heading tag="h6">Heading 6</Heading>
 </Playground>
 
-## Styling headers
+## Styling headings
 
 <Playground>
-  <Header color="primary" lineHeight="1" fontWeight="100">Header 1</Header>
-  <Header textAlign="center">Center aligned header</Header>
+  <Heading color="primary" lineHeight="1" fontWeight="100">Heading 1</Heading>
+  <Heading textAlign="center">Center aligned heading</Heading>
 </Playground>
 
 ### PropTypes
@@ -76,5 +76,5 @@ Use the `tag` prop to change from the default `h1` element to `h2`, `h3`, `h4`, 
 }
 ```
 
-- [Default theme](https://github.com/raster-foundry/blasterjs/tree/master/packages/core/theme/components/header)
+- [Default theme](https://github.com/raster-foundry/blasterjs/tree/master/packages/core/theme/components/heading)
 - [Learn more about themeing](#)
