+++ packages/core/theme/components/heading/index.js
